move zip handler to utils
