+++ src/allotropy/parsers/utils/zip_handler.py
